Remove component import from specs
As of playwright version 1.23.0, importing from tsx files seems to
be problematic. Related issues:

* https://github.com/microsoft/playwright/issues/15076
* https://github.com/microsoft/playwright/issues/11720

Since the import was only a type definition that isn't adding much
value (e.g. if the response structure drifts from the expected tests
will fail anyways), this change removes it to resolve the following
failure message otherwise coming from ContentLoader.tsx:

    SyntaxError: Unexpected token '<'

An alternative would have been to move the type definition to a .ts
file, but this seemed to be the lesser evil since it's not really
used anywhere else.