Explicitly define output filename for prod builds
Per https://github.com/vitejs/vite/pull/10927, there was a change
made as part of vite 4 to align default build file names with the
default from rollup. While this is mostly superficial for us, since
the change causes a slight difference in the production filename
artifact I'm explicitly adding a configuration for it.