Fix jest tests failures related to mathlive
I'm not sure whether this is the best fix, and I'm specifically unclear
as to why the moduleNameMapper entry in jest.config.ts is necessary to
avoid errors, but this commit at least is one approach for fixing
unit test errors I realized were occurring. I believe the roof of the
problem is jsdom does not emulate ResizeObserver. Here we mock that
out using a new jest.setup.ts so existing tests work.