Update tests and jest config for msw v2
This change includes some pretty hacky stuff to enable updating to
v2 of the msw library used by some of our tests. The documented
migration steps didn't work as adverised so I resorted to applying
work arounds discussed in GitHub issues. It's likely some of this may
need to get revisited in the future as the approach of the custom jest
resolver in particular feels pretty brittle.